finance/requirements.txt
azure-storage-blob>=12.0.0
azure-ai-formrecognizer>=3.2.0
azure-search-documents>=11.4.0b8
azure-core>=1.26.0
openai>=1.12.0
python-dotenv>=0.19.0
streamlit>=1.25.0
